All features completed

+++ app/Http/Controllers/API/v1/Orders/OrderController.php
@@ -29,59 +29,53 @@ class OrderController extends ApiController {
     }
 
     public function list(Request $request, Workflow $workflow) {
-        // $this->authorize('viewAny', Workflow::class);
+        $this->authorize('viewAny', Order::class);
         $orders = $this->orderRepository->list($workflow, $request->all(), true, true);
         return $this->responseWithData(200, $orders);
     }
 
     public function create(CreateRequest $request, Workflow $workflow) {
-        // $this->authorize('create', Workflow::class);
+        $this->authorize('create', Order::class);
         $this->orderRepository->create($workflow, $request->all(), $request->files->all());
         return $this->responseWithMessage(201, 'Order created.');
     }
 
+    public function update(UpdateRequest $request, Workflow $workflow, $id) {
+        $order = $this->orderRepository->find($workflow, $id);
+        if ($order == null) 
+            throw new ModelNotFoundException();
+
+        $this->authorize('update', $order);
+        $this->orderRepository->update($workflow, $order, $request->all(), $request->files->all());
+        return $this->responseWithMessage(200, 'Order updated.');
+    }
+
     public function details(Workflow $workflow, $id) {
-        // $this->authorize('view', $workflow);
         $order = $this->orderRepository->find($workflow, $id);
         if ($order == null) 
             throw new ModelNotFoundException();
         
+        $this->authorize('view', $order);
         return $this->responseWithData(200, $order);
     }
 
     public function updateProcess(UpdateProcessRequest $request, Workflow $workflow, $id) {
-        $this->orderRepository->updateProcess($workflow, $id, $request->all());
+        $order = $this->orderRepository->find($workflow, $id);
+        if ($order == null) 
+            throw new ModelNotFoundException();
+
+        $this->orderRepository->updateProcess($workflow, $order, $request->all());
         return $this->responseWithMessage(200, 'Order Process updated.');
     }
 
     public function delete(Workflow $workflow, $id) {
+        $order = $this->orderRepository->find($workflow, $id);
+        if ($order == null) 
+            throw new ModelNotFoundException();
+        
+        $this->authorize('delete', $order);
         $this->orderRepository->delete($workflow, $id);
         return $this->responseWithMessage(200, 'Order Deleted.');
     }
 
-    // public function update(UpdateRequest $request, Workflow $workflow) {
-    //     $this->authorize('update', $workflow);
-    //     $this->workflowRepository->update($workflow, $request->all());
-    //     return $this->responseWithMessage(200, 'Workflow updated.');
-    // }
-
-    // public function activate(Request $request, Workflow $workflow) {
-    //     $this->authorize('update', $workflow);
-    //     $data['status'] = Workflow::STATUS_ACTIVE;
-    //     $this->workflowRepository->update_status($workflow, $data);
-    //     return $this->responseWithMessage(200, 'Workflow activated.');
-    // }
-
-    // public function deactivate(Request $request, Workflow $workflow) {
-    //     $this->authorize('update', $workflow);
-    //     $data['status'] = Workflow::STATUS_INACTIVE;
-    //     $this->workflowRepository->update_status($workflow, $data);
-    //     return $this->responseWithMessage(200, 'Workflow deactivated.');
-    // }
-
-    // public function delete(Workflow $workflow) {
-    //     $this->authorize('delete', $workflow);
-    //     $this->workflowRepository->delete($workflow);
-    //     return $this->responseWithMessage(200, 'Workflow deleted.');
-    // }
 }
